Fix configuration
The d89b8e5 commit introduces a regression in the original shippo
implementation, with the introduction of a `config` object within the
`config` module.
The original documentation was mentioning configuring the package with
`shippo.config.<key> = <value`. It wasn't working anymore.
This changeset hasn't been published to pypi, but the master branch was
containing it.

Changeset:
- renames the `config` module to `configuration` to remove the ambiguity.
- The shippo package now imports the config object from the configuration
  module, fixing the expected behavior.
- Test has been added to prevent future regression.